padroes de projeto revisão

+++ src/main/java/br/ufjf/dcc193/projeto_ufjf/Repositorio/AreaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufjf.dcc193.projeto_ufjf.Modelo.Area;
 
-public interface AreaRepositorio extends JpaRepository<Area, Integer>{
+public interface AreaRepositorio extends JpaRepository<Area, Long>{
 
 }
