+++ src/main/java/br/ufjf/dcc193/projeto_ufjf/Repositorio/AvaliadorRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufjf.dcc193.projeto_ufjf.Modelo.Avaliador;
 
-public interface AvaliadorRepositorio extends JpaRepository<Avaliador, Integer>{
+public interface AvaliadorRepositorio extends JpaRepository<Avaliador, Long>{
 
 }
