+++ src/main/java/br/ufjf/dcc193/projeto_ufjf/Repositorio/RevisaoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufjf.dcc193.projeto_ufjf.Modelo.Revisao;
 
-public interface RevisaoRepositorio extends JpaRepository<Revisao, Integer>{
+public interface RevisaoRepositorio extends JpaRepository<Revisao, Long>{
 
 }
