+++ src/main/java/br/ufjf/dcc193/projeto_ufjf/Repositorio/TrabalhoRepositorio.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufjf.dcc193.projeto_ufjf.Modelo.Trabalho;
 
-public interface TrabalhoRepositorio extends JpaRepository<Trabalho, Integer>{
+public interface TrabalhoRepositorio extends JpaRepository<Trabalho, Long>{
 
 
 }
